# gateway/docs/openapi.yaml
openapi: 3.1.0
info:
  title: WaaV Gateway API
  description: Real-time voice processing server with Speech-to-Text (STT) and Text-to-Speech (TTS) services
  contact:
    name: WaaV
    url: https://api.waav.ai
  license:
    name: Apache-2.0
    identifier: Apache-2.0
  version: 0.1.0
servers:
- url: https://api.waav.ai
  description: Production API
- url: http://localhost:3001
  description: Local development
paths:
...
  /speak:
    post:
      tags:
      - tts
      summary: Handler for the /speak endpoint
      operationId: speak_handler
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SpeakRequest'
        required: true
      responses:
        '200':
          description: Audio generated successfully
          headers:
            x-audio-format:
              schema:
                type: string
              description: Audio format (linear16, mp3, etc.)
            x-sample-rate:
              schema:
                type: integer
                format: int32
                minimum: 0
              description: Sample rate in Hz
          content:
            audio/pcm: {}
        '400':
          description: Invalid request (empty text)
        '500':
          description: TTS synthesis failed
      security:
      - bearer_auth: []
...
components:
  schemas:
    Age:
      type: string
      description: |-
        Canonical voice age band. ElevenLabs uses exactly `young`/`middle_aged`/`old`;
        Azure roles are `YoungAdult*`/`OlderAdult*`/`Senior*` (Senior folds into `Old`).
      enum:
      - young
      - middle_aged
      - old
...
    DeliveryStyle:
      type: string
      description: |-
        Delivery style modifiers for speech synthesis.

        These styles can be combined with emotions to further customize
        how the speech is delivered. They affect pacing, volume, and
        prosodic patterns.

        # Example

        ```rust
        use waav_gateway::core::emotion::DeliveryStyle;

        let style = DeliveryStyle::Whispered;
        assert_eq!(style.to_string(), "whispered");
        ```
      enum:
      - normal
      - whispered
      - shouted
      - rushed
      - measured
      - monotone
      - expressive
      - professional
      - casual
      - storytelling
      - loud
      - formal
      - newscast
      - newscast_casual
      - newscast_formal
      - customer_service
      - assistant
      - chat
      - advertisement_upbeat
      - sports_commentary
      - sports_commentary_excited
      - documentary_narration
      - narration_professional
      - narration_relaxed
      - poetry_reading
      - lyrical
      - gentle
    Emotion:
      type: string
      description: |-
        Standardized emotions supported across TTS providers.

        These emotions represent common emotional states that can be expressed
        through speech synthesis. Different providers support different subsets
        of these emotions, and the emotion mapper handles the translation.

        # Provider Support Matrix

        | Emotion | Hume | ElevenLabs | Azure | OpenAI |
        |---------|------|------------|-------|--------|
        | Neutral | Yes  | Yes (default) | Yes | Yes |
        | Happy   | Yes  | Yes (low stability) | Yes (cheerful) | Yes |
        | Sad     | Yes  | Yes (high stability) | Yes | Yes |
        | Angry   | Yes  | Yes (low stability) | Yes | Yes |
        | Fearful | Yes  | Partial | Yes (terrified) | Partial |
        | Surprised | Yes | Partial | Yes | Partial |
        | Excited | Yes | Yes | Yes (excited) | Yes |
        | Calm    | Yes | Yes (high stability) | Yes (calm) | Yes |
        | Sarcastic | Yes | Partial | Partial | Yes |

        # Example

        ```rust
        use waav_gateway::core::emotion::Emotion;

        let emotion = Emotion::Happy;
        assert_eq!(emotion.to_string(), "happy");
        assert!(emotion.is_widely_supported());
        ```
      enum:
      - neutral
      - happy
      - sad
      - angry
      - fearful
      - surprised
      - disgusted
      - excited
      - calm
      - anxious
      - confident
      - confused
      - empathetic
      - sarcastic
      - hopeful
      - disappointed
      - curious
      - grateful
      - proud
      - embarrassed
      - content
      - bored
      - amazed
      - amused
      - affectionate
      - intrigued
      - flirtatious
      - frustrated
      - annoyed
      - determined
      - reassuring
      - sympathetic
      - nostalgic
      - serene
      - terrified
      - ecstatic
      - skeptical
      - relieved
      - panicked
      - concerned
      - apologetic
      - resigned
      - wistful
      - contemplative
...
    EmotionIntensity:
      oneOf:
      - type: number
        format: float
        description: Numeric intensity (0.0 to 1.0)
      - $ref: '#/components/schemas/IntensityLevel'
        description: Named intensity level
      description: |-
        Unified emotion intensity that supports both numeric and named levels.

        This allows users to specify intensity either as a numeric value (0.0-1.0)
        or as a named level (low, medium, high) for convenience.

        # Example

        ```rust
        use waav_gateway::core::emotion::{EmotionIntensity, IntensityLevel};

        // Numeric intensity
        let intensity = EmotionIntensity::Numeric(0.8);
        assert_eq!(intensity.as_f32(), 0.8);

        // Named intensity
        let intensity = EmotionIntensity::Named(IntensityLevel::High);
        assert_eq!(intensity.as_f32(), 1.0);
        ```
    Gender:
      type: string
      description: |-
        Canonical voice gender. Providers spell this `Female`/`Male`/`Neutral` (Google
        `ssml_gender`), `"female"`/`"male"` (ElevenLabs labels), etc.
      enum:
      - male
      - female
      - neutral
...
    IntensityLevel:
      type: string
      description: |-
        Named intensity levels for emotion expression.

        These map to numeric values for providers that support intensity:
        - Low: 0.3
        - Medium: 0.6
        - High: 1.0

        Providers can also accept raw numeric intensity (0.0 to 1.0).
      enum:
      - low
      - medium
      - high
...
    Pronunciation:
      type: object
      description: Pronunciation replacement configuration
      required:
      - word
      - pronunciation
      properties:
        pronunciation:
          type: string
          description: Pronunciation to use instead
          example: A P I
        word:
          type: string
          description: Word to replace
          example: API
    ProviderExtras:
      type: object
      description: Open provider-specific passthrough (string→JSON value map). Keys not modeled by the typed feature vocabulary are forwarded verbatim.
      additionalProperties: true
...
    SpeakRequest:
      type: object
      description: Request body for the speak endpoint
      required:
      - text
      - tts_config
      properties:
        text:
          type: string
          description: The text to synthesize
          example: Hello, world!
        tts_config:
          $ref: '#/components/schemas/TTSWebSocketConfig'
          description: TTS configuration (without API key)
...
    TTSWebSocketConfig:
      type: object
      description: TTS configuration for WebSocket messages (with optional API key)
      required:
      - provider
      - model
      properties:
        api_key:
          type:
          - string
          - 'null'
          description: Optional API key for this provider (overrides server config)
        audio_format:
          type:
          - string
          - 'null'
          description: Audio format preference
          example: linear16
        audio_out_chunk_ms:
          type:
          - integer
          - 'null'
          format: int32
          description: |-
            WS egress re-framing (E-G2): when set, PCM TTS audio is sliced into
            chunks of this many milliseconds before the WebSocket send, so a
            barge-in clear truncates within ONE chunk instead of letting a whole
            synthesis frame play out client-side. Omitted = verbatim frames
            (legacy). LiveKit egress already frames at 10ms internally.
          example: 20
          minimum: 0
        audio_out_codec:
          type:
          - string
          - 'null'
          description: |-
            D8 downlink TRANSPORT codec for TTS egress: `linear16` (default — raw PCM16 frames) | `opus`
            (the gateway encodes each normalized PCM16 chunk to one opus packet per WS binary frame; the
            client decodes before playout). Encodes from the C-G5-normalized PCM16 stream and reuses
            `audio_out_chunk_ms` as the opus frame size (default 20ms) + `client_playback_rate` as the opus
            rate (constrained to {8,12,16,24,48}kHz; default 48000 when opus). A non-PCM provider stream
            (mp3/ogg) falls back to passthrough + warns. Negotiated: an `opus` request on a gateway without
            the `opus-codec` feature degrades to `linear16` (`config_warning`) and the effective codec is
            echoed in `ready`.
          example: opus
        client_playback_rate:
          type:
          - integer
          - 'null'
          format: int32
          description: |-
            Client playback rate (C-G5): when set, the gateway resamples PCM TTS
            egress from the provider's rate to THIS rate before delivery (WS
            binary and LiveKit), so clients with a fixed-rate audio sink never
            resample provider-side or client-side. Identity (provider already at
            this rate) is zero-cost; non-PCM formats pass through unchanged.
          example: 48000
          minimum: 0
        connection_timeout:
          type:
          - integer
          - 'null'
          format: int64
          description: Connection timeout in seconds
          example: 30
          minimum: 0
        delivery_style:
          oneOf:
          - type: 'null'
          - $ref: '#/components/schemas/DeliveryStyle'
            description: |-
              Delivery style modifier for speech.

              Affects pacing and prosody independently of emotion.
              Examples: "whispered", "shouted", "rushed", "measured", "expressive"
        emotion:
          oneOf:
          - type: 'null'
          - $ref: '#/components/schemas/Emotion'
            description: |-
              Emotion to express in speech synthesis.

              Supported by: Hume AI (all), ElevenLabs (core set), Azure (SSML styles).
              Providers without emotion support will synthesize speech normally with a warning.

              Examples: "happy", "sad", "angry", "excited", "calm", "sarcastic"
        emotion_description:
          type:
          - string
          - 'null'
          description: |-
            Free-form emotion description (for providers like Hume AI).

            When provided, this takes precedence over the `emotion` field for Hume AI.
            Maximum 100 characters.

            Examples: "warm, friendly, inviting", "whispered fearfully", "sarcastic, dry"
          example: warm, friendly, inviting
        emotion_intensity:
          oneOf:
          - type: 'null'
          - $ref: '#/components/schemas/EmotionIntensity'
            description: |-
              Intensity of the emotion (0.0 to 1.0, or "low"/"medium"/"high").

              - 0.0-0.3: Subtle expression
              - 0.4-0.7: Moderate expression (default: 0.6)
              - 0.8-1.0: Strong expression

              Alternatively, use named levels: "low" (0.3), "medium" (0.6), "high" (1.0)
        extras:
          $ref: '#/components/schemas/ProviderExtras'
          description: Open, typed passthrough for provider-specific parameters not modeled by `features`.
        features:
          $ref: '#/components/schemas/TtsFeatures'
          description: |-
            Canonical, provider-agnostic advanced TTS features (voice settings, instructions, SSML,
            streaming, …). Defaults to all-unset so existing clients are unaffected; carried across the
            dispatch boundary via [`Self::to_standard_tts`] so providers with a `from_standard` mapping
            (Deepgram first) honor them END-TO-END (W1 keystone, closing BRUTAL_REVIEW.md S1/S5).
        model:
          type: string
          description: Model to use for TTS
          example: aura-asteria-en
        pronunciations:
          type: array
          items:
            $ref: '#/components/schemas/Pronunciation'
          description: Pronunciation replacements to apply before TTS
        provider:
          type: string
          description: Provider name (e.g., "deepgram", "hume", "elevenlabs")
          example: deepgram
        request_timeout:
          type:
          - integer
          - 'null'
          format: int64
          description: Request timeout in seconds
          example: 60
          minimum: 0
        sample_rate:
          type:
          - integer
          - 'null'
          format: int32
          description: Sample rate preference
          example: 24000
          minimum: 0
        speaking_rate:
          type:
          - number
          - 'null'
          format: float
          description: Speaking rate (0.25 to 4.0, 1.0 is normal)
          example: 1.0
        voice_descriptor:
          oneOf:
          - type: 'null'
          - $ref: '#/components/schemas/VoiceDescriptor'
            description: |-
              Canonical voice DESCRIPTOR (P4): `{gender, locale/accent, style, age,
              name_hint}` resolved SERVER-SIDE to a provider `voice_id` over the `/voices`
              catalog when no raw `voice_id` is given. No match → provider default +
              `config_warning` (never a 400). The resolved id is echoed back to the client.
        voice_id:
          type:
          - string
          - 'null'
          description: |-
            Voice ID or name to use for synthesis.

            ESCAPE HATCH: when set, this raw provider id is used VERBATIM and the
            [`Self::voice_descriptor`] resolution is skipped.
          example: aura-asteria-en
...
    VoiceDescriptor:
      type: object
      description: |-
        A canonical, provider-agnostic description of a desired voice. Resolved
        SERVER-SIDE to a concrete provider `voice_id` over the `/voices` catalog.

        All fields are optional; the resolver scores candidates on whichever are set.
        The raw `TTSConfig.voice_id` (when present) ALWAYS wins — this descriptor only
        drives resolution when no explicit `voice_id` was supplied.

        # Example

        ```rust
        use waav_gateway::core::voice::{VoiceDescriptor, Gender};

        let d = VoiceDescriptor {
            gender: Some(Gender::Female),
            locale: Some("en-US".to_string()),
            style: Some("warm".to_string()),
            ..Default::default()
        };
        assert!(d.is_set());
        ```
      properties:
        accent:
          type:
          - string
          - 'null'
          description: Free accent string when no `locale` is given (`american`, `british`, …).
          example: american
        age:
          oneOf:
          - type: 'null'
          - $ref: '#/components/schemas/Age'
            description: Desired age band.
        gender:
          oneOf:
          - type: 'null'
          - $ref: '#/components/schemas/Gender'
            description: Desired gender.
        locale:
          type:
          - string
          - 'null'
          description: |-
            BCP-47 locale (preferred): `en-US`, `en-GB`, `hi-IN`, … Used for exact-locale
            and language-family scoring; the region also yields an accent string.
          example: en-US
        name_hint:
          type:
          - string
          - 'null'
          description: |-
            Optional preferred voice name (e.g. `asteria`, `jenny`). A substring match
            against the voice name/id gives it the strongest score.
          example: asteria
        style:
          type:
          - string
          - 'null'
          description: |-
            Free descriptive timbre/style matched against provider metadata
            (`warm`, `bright`, `deep`, `calm`, `confident`, `energetic`, …).
          example: warm
  securitySchemes:
    bearer_auth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from the authentication service. Required when AUTH_REQUIRED is enabled.
...